requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.41.0
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.8.0rc2
distro==1.9.0
Django==4.2.13
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
Faker==26.0.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
inflection==0.5.1
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.4.3
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
